added initial media query check in useMediaQuery hook
The useMediaQuery hook now includes an initial check to set the initial state of the media query based on the viewport status. This ensures that the hook correctly reflects the initial state of the media query when the component first renders, avoiding potential bugs on page load or refresh. Bug which is solved by this solution was that navbar was always shown for responsive on first render and page reload.